src/main/webapp/app/entities/serviciosmedicos/tramite/delete/tramite-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { TramiteService } from '../service/tramite.service';

import { TramiteDeleteDialogComponent } from './tramite-delete-dialog.component';

describe('Tramite Management Delete Component', () => {
  let comp: TramiteDeleteDialogComponent;
  let fixture: ComponentFixture<TramiteDeleteDialogComponent>;
  let service: TramiteService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, TramiteDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(TramiteDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(TramiteDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(TramiteService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
